requirements.txt
qdrant-client
langchain
langchain_community
sentence-transformers
pypdf
torch
fpdf
groq
python-dotenv